add type:ignores

+++ docs/epytext_demo/demo_epytext_module.py
@@ -28,8 +28,8 @@ This is also a constant, but annotated with typing.Final.
 """
 
 @deprecated(Version("demo", "NEXT", 0, 0), replacement=math.prod)
-def demo_product_deprecated(x, y) -> float:
-    return x * y
+def demo_product_deprecated(x, y) -> float: # type: ignore
+    return float(x * y)
 
 def demo_fields_docstring_arguments(m, b):  # type: ignore
     """
